Update to form admin de-dupe code and shifting enable toggling to action
Several duplicated modelAdmin entries moved into mixin.
Added search of name and callable fields to admin.
Instead of enabled being a list_editable field there's now a form action to enable or disable all in a selection. This is useful when paired with the list_filter and search fields allowing for bulk enabling / disabling of jobs.